Move NvAPI in the plugin folder (CST-1465)
