+++ GfxPluginQuadroSync/External/NvAPI/License.txt
