+++ GfxPluginQuadroSync/External/NvAPI/NvApiDriverSettings.c
