+++ GfxPluginQuadroSync/External/NvAPI/NvApiDriverSettings.h
