+++ GfxPluginQuadroSync/External/NvAPI/amd64/nvapi64.lib
