+++ GfxPluginQuadroSync/External/NvAPI/nvHLSLExtns.h
