+++ GfxPluginQuadroSync/External/NvAPI/nvHLSLExtnsInternal.h
