+++ GfxPluginQuadroSync/External/NvAPI/nvapi.h
