+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_common.h
