+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_d3dext.h
