+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_salend.h
