+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_salstart.h
