+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_sli.h
