+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_stereo.h
