+++ GfxPluginQuadroSync/External/NvAPI/nvapi_lite_surround.h
