+++ GfxPluginQuadroSync/External/NvAPI/x86/nvapi.lib
